requirements.txt
argcomplete==3.6.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.37.19
botocore==1.37.19
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfn-flip==1.3.0
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
durationpy==0.9
google-auth==2.38.0
hjson==3.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==3.0.2
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.1.0
placebo==0.9.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.4
service-identity==24.2.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
text-unidecode==1.3
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
troposphere==4.9.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zappa==0.59.0
zope.interface==7.2
# added these for oAuth compatibility
social-auth-app-django==5.4.3
google-auth==2.38.0
google-generativeai==0.1.0rc1 #bonnie
coverage==7.8.0